Add files via upload
To collect the data is not enough to run the successful business or any company. The most important thing to get right analysis of that collected data. And that's what i gave to Vrinda Store which helps to grow their business easily and known where they focused or approach a particular type of Customers, on the based of previous Statistics.